Update version 1.1
Updated README.MD as suggested in #3
Fixed security issue of #4 as suggested in #5.
The script now stores username and password in a default configuration file ( ~/.config/polito_dl/auth on Linux/MacOS and %appdata%\polito_dl\auth on Windows).
On Linux/MacOS it also checks that proper file permissions are set.

Closes #3, closes #4 and  closes #5.